macro_synteny_paf/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
aiohttp==3.8.3
    # via
    #   aiohttp-cors
    #   macro-synteny-paf (setup.py)
aiohttp-cors==0.7.0
    # via macro-synteny-paf (setup.py)
aiosignal==1.3.1
    # via aiohttp
async-timeout==4.0.2
    # via
    #   aiohttp
    #   redis
attrs==22.2.0
    # via aiohttp
charset-normalizer==2.1.1
    # via aiohttp
frozenlist==1.3.3
    # via
    #   aiohttp
    #   aiosignal
grpcio==1.51.1
    # via
    #   grpcio-tools
    #   macro-synteny-paf (setup.py)
grpcio-tools==1.51.1
    # via macro-synteny-paf (setup.py)
idna==3.4
    # via yarl
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
protobuf==4.21.12
    # via grpcio-tools
redis==4.4.2
    # via macro-synteny-paf (setup.py)
uvloop==0.17.0
    # via macro-synteny-paf (setup.py)
yarl==1.8.2
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
